\name{mixl-package}
\alias{mixl-package}
\alias{mixl}
\docType{package}
\title{
  Estimate mixed multinomial logit models
}
\description{
  Estimate mixed multinomial logit models using (simulated) maximum likelihood estimation. 
  The package supports standard mnl, mixed-logit and hybrid choice. Using compilation to C++,
  model estimation is significantly faster than in native R code.
}
\details{
  This section should provide a more detailed overview of how to use the
  package, including the most important functions.
}
\author{
Joe Molloy <joseph.molloy@ivt.baug.ethz.ch>.
}
\references{
  This optional section can contain literature or other references for
  background information.
}
\keyword{ package }
\seealso{
  Optional links to other man pages
}
\examples{
  \dontrun{
     ## Optional simple examples of the most important functions
     ## These can be in \dontrun{} and \donttest{} blocks.   
  }
}
